Unicast packet handling from host on cold boot
Summary:
There are 2 variations of sendL3Packet that we support
i) With a interface ID (must for link local handling). Tun interfaces always
use this.
ii) Without a interface ID

Tests for COLD boot handling of i) were added in D15006096. So this is
about ii), specifically unicast packets for case ii. In this case, 2
things happen
- Unicast packet is sent to the ASIC
- Neighbor resolution is attempted, but since no next hops can be looked up
we dont send out any NDP/ARP packets.
So just assert for sendPacketSwitchedAsync to be called once on MockHwSwitch

Separate diffs coming for multicast and link local destinations

Reviewed By: saifhhasan

Differential Revision: D15009613

fbshipit-source-id: a888de571be9844d60e89c9665136d6e5bb3f677