First tests for packet handling when SwSwitch state is set to cold boot state
Summary:
As part of our effort to get packet handling be available early,
we want to start handling packets as soon as ASIC is initialized vs
waiting for config application. Post warm boot this is not a problem
since we recover to (nearly) the same SwitchState as we went down with.
However post cold boot, we would recover to the cold boot state which
maps
- All front panel ports disabled
- All ports in a single VLAN

Ports being disabled shields us from front panel ports but not from
packets coming from host cpu via the tun interfaces. Our current code
though has guards to protect against that - viz TunIntf sends a packet
with the Tun interface's interface ID, which will not exist in cold boot
state. sendL3Packet however drops the packet on the floor if it can't
find the interface ID. The code above just exercises this code path
and asserts this behavior in perpetuity.

Reviewed By: saifhhasan

Differential Revision: D15006096

fbshipit-source-id: 5e5591a14642bd571fb05a781f6864ca8f0eed82