Multicast packet handling from host on cold boot
Summary:
There are 2 variations of sendL3Packet that we support
i) With a interface ID (must for link local handling). Tun interfaces always
use this.
ii) Without a interface ID

Tests for COLD boot handling of i) were added in D15006096. So this is
about ii), specifically multicast packets for case ii. For such packets
we just send the out to be switched by h/w. So just assert for sendPacketSwitchedAsync
being called.
Unicast packet handling was tested in previous diff. Link local packet
testing is in next one

Reviewed By: saifhhasan

Differential Revision: D15009616

fbshipit-source-id: 09bd337e954854e6df62dfd01f7a453d38219b0b