Link local packet handling from host on cold boot
Summary:
There are 2 variations of sendL3Packet that we support
i) With a interface ID (must for link local handling). Tun interfaces always
use this.
ii) Without a interface ID

Tests for COLD boot handling of i) were added in D15006096. So this is
about ii), specifically link local packets for case ii. For such packets
since no VLAN was provided and this is a link local packet. We drop packet
on the floor.

Reviewed By: saifhhasan

Differential Revision: D15009881

fbshipit-source-id: 11a81de08483c25889d57efbf31ceb129d05fb11