accept pop and lookup without interface
Summary:
integeration with open-r revealed that open-r does not set interface id for pop and lookup forwarding action.

this action defines forwarding behavior on 'exposed header' after popping the incoming label and so the interface id (and next hop address) is not required.

also this action must be mentioned with only one next hop. so next hop set must be of size 1.

enforcing this in code.

Reviewed By: saifhhasan

Differential Revision: D15092109

fbshipit-source-id: 3d01b3042011a1754efdc53daa02de07a4e02ef9