move rx peaking filter out of config
Summary:
Recovering this field is quite unreliable and we use as one
indication of whether we need to reprogram a phy. This setting is
always set on system side lanes right now to the same value, never on
line side lanes. Because of that, it is easier to just set as part of
bringup process and skip setting in config.

Differential Revision: D15100063

fbshipit-source-id: 123292e8a09d308cda5ea131e46ebc82db0de5ba