rework platform_config a bit to support per-lane ASIC TAPs
Summary:
HW team + accton really optimized tx for 100g ports on
minipack and want fine grained control of tx taps for each lane on the
TH3. Previous thrift model would only work if the pre-emphasis was the
same for every lane in a logical port. This is not sufficient for the
settings we want, so just reuse the external phy side setting struct.

I'll rename to just "phy" down the line.

Reviewed By: xzhu70

Differential Revision: D15117149

fbshipit-source-id: 4e92bfc94403f08f5c45d3646e49ce215d14fac8