normalized next hops in route nexthop entry
Summary:
bringing the logic of normalizing next hops in RouteNextHopEntry

also moved ecmp_width gflag in RouteNextHopEntry / SwSwitch's state.

this will be reused in supporting ECMP/UCMP for MPLS routes as well.

no other code changes in the existing logic of normalizing next hops, this diff merely pulls the code out of BcmLayer and makes it available in SwSwitch layer in an appropriate class.

Differential Revision: D15155034

fbshipit-source-id: b30c23091a95939ef89446773e9fb0f2e4f46e65