sai ping hack cleanup -- process delta fixes
Summary:
We were taking the shared ptrs from the delta by value
instead of const reference which is not idiomatic to the delta
processing code. I _believe_ it even caused some kind of racy
crashes, but haven't fully proved it. Either way, this is an
easy change to get in line with the rest of delta processing out
there. Also actually include route deltas in top level stateChanged

Differential Revision: D15261247

fbshipit-source-id: 43dddf08c87129fb7ed5c34bbf6648165e7bdfe0