no need to replace BcmPort MonotonicCounter if name matches
Summary:
If I call reinitPortStats twice without changing the name,
there is no need to swap these counters.

This becomes more significant as I add logic to skip port counters on
disabled ports. I call reinitPortStats when we enable the port. In the
case that we flapped a port, there is no need to reset all the
counters like this. We have a known race condition in this area and
avoiding unnecessary operations like this seems desirable.

Reviewed By: shrikhare

Differential Revision: D15344425

fbshipit-source-id: 6e47f47174605ca2f1b2be4d1c5b8156722d85ed