no bcm stats for disabled ports
Summary:
This skips updateStats for any BcmPorts that are disabled. We
now have a boolean that determines whether or not we should collect
stats. We set this when we enable/disable a port, or when we recover a
port post-warmboot.

I considered using isEnabled() calls directly, but this introduces a
lot of redundant sdk calls so this approach seemed better.

Differential Revision: D15152121

fbshipit-source-id: f1c689a79cc0755a79fe6afbded4dcceeb2699d0