Apply config to stand-alone RIB and FIB
Summary:
This extends RoutingInformationBase with
```RoutingInformationBase::reconfigure()```. This method takes as input those
routes which are defined by configuration (namely static routes, interface
routes, and link-local routes), inserts them into the Routing Information Base,
and updates the Forwarding Information Base with those which are IP resolved.

Most of this logic is factored out into a ```rib::ConfigApplier``` functor.

To faciliate the transition to a stand-alone RIB, the set of interace routes is
still generated in ApplyThriftConfig and only passed on to ConfigApplier for
processing.  That is, the logic determining the set of interfaces routes remains
intact from combined RIB implementation to the stand-alone RIB. Once the FBOSS
agent has been completely cut-over to the stand-alone RIB, this logic should be
moved into ```rib::ConfigApplier```.

Reviewed By: tfangit

Differential Revision: D13992939

fbshipit-source-id: c3425e8227540c558b6f30892091de242df15893