Added support for LLDP based LED colours in YAMP
Summary:
Currently only Minipack has support for changing the LED colour
based on LLDP data not matching the expected LLDP in the config. This diff adds
this support to YAMP also.

This also fixes a bug in the initial implementation. We now check the system_name and port fields in LLDP. Before we were checking system_name and port_desc fields which does not make sense as port_desc is more variable. port_desc also describes the expected remote end of the link (not the interface the lldp packet was sent from)

Differential Revision: D15334600

fbshipit-source-id: e686bc9a9bfbe94aeea5081c49e20cdc69c9fbb9