Add module to track restart timing
Summary:
This adds a module to the agent for tracking times of events
across reboot and then exporting them as fb303 counters. It exports
the time per each stage as well as the total time the process took.

The logic is aware of which type of restart we performed and will
export counters w/ either warm_boot/cold_boot prefixes based on that.

The tricky part here is that we need to save some state across process
invocations. We do this by dumping timestamps to files on shutdown and
then recovering them on subsequent bootup.

Reviewed By: shrikhare

Differential Revision: D15522955

fbshipit-source-id: a56429e5b2ec402b0b029e6ccdd1fd982a276265