Update Broadcom ASIC with FIB changes
Summary:
This updates BcmSwitch to unconditionally consume routes from the
SwitchState ForwardingInformationBase complex.

Note that routes are consumed from both the ForwardingInformationBase and the
RouteTable.

I deliberated between this approach and propogating the
FLAGS_enable_standalone_rib from SwSwitch down to BcmSwitch to pick which
SwitchState hierarchy to consume routes from. I chose this approach because it
allows for code paths to be bifurcated at fewer callsites, and with the
soundness assertion, it is no less safe.

Reviewed By: phshaikh

Differential Revision: D12808050

fbshipit-source-id: 819dd620e1b1fa0ba037f8e7292b8ca645a615e5