fix publishing of ecmp groups used statistic
Summary:
As reference counting of egresses went out of BcmHostTable, a stat of "l3_ecmp_groups_used" was not available in BcmHostTable anymore

The counter which kept track of ecmp egresses also was eliminated

Restoring this stat as a counter in BcmEcmpEgress class.

BcmEcmpEgress maintains a counter of active objects and providing API to increment, decrement, and extract the counter.

Differential Revision: D15485110

fbshipit-source-id: 35cbe8b168a5190dc26b789f3ecdd6a9c1fb7d6b