make SaiSwitch thread safe
Summary:
We want the public API of SaiSwitch to be thread safe. However, the
API of HwSwitch was designed with one API possibly calling another in mind.
To distinguish between those cases, we use the basic tactic of a private
"locked" version of each API. External calls will use the public API and
take the lock, internal calls will use the private API and assume the lock is
held. Hopefully this can also scale to more granular locking as needed.

Differential Revision: D15611011

fbshipit-source-id: 554731596694b10e2d4b643303aaf71d010c14e8