Add support for LLDP based LED control to Wedge100
Summary:
Add support for LLDP based LED control to Wedge100 devices.
These devices already support diffent LED colours based on the QSFP type.
This code should work the same way as minipack in that the LED will be
changed to YELLOW if we detect a conflict between the LLDP information
on the wire and LLDP in the config

Reviewed By: aeckert

Differential Revision: D15694147

fbshipit-source-id: 951c6b9eab2b7b38a1ccc3167b1c150836bcd1a6