reprogram mpls next hops on neighbor update
Summary:
In response to changes in neighbor entry, MPLS next hop to that neighbor must be updated.

This is done in this diff.

If neighbor entry is reachable, MPLS nexthop is programmed to the port/trunk over which next hop is reachable

If neighbor entry is pending, MPLS nexthop is programmed to CPU

If neighbor entry is unknown, MPLS nexthop is programmed to drop.

For L3 next hops, no action is required - as L3 next hops are nothing but Host Entry (BCMHost) which are already updated as a part of neighbor entry changes.

Differential Revision: D15321803

fbshipit-source-id: c82a8ce8d483c868772f136827df1062f1834a49