Don't drop h/w out of sync updates
Summary:
When we get a update that cannot be applied to hw, we
requeue it back into pending updates and then service it again when
i) We get new state updates
ii) Periodically when we our timer to handle pending updates fires.

For ii) we have a optimization that if the h/w out of sync is the only
update in the queue. Don't try to reapply it to h/w. However there
was a serious bug here, where we had already pulled this update
out of the pendingUpdatesList but then when we skipped applying this
update we never queued it back. This meant we were now *dropping* the
offending update on the floor.
Fix is to move this check earlier and forgo pulling things out
of pendingUpdatesList_ if hw_out_of_sync is the only update here.

TODO: Identifying hw_out_of_sync by name is pretty janky. Make this
a property of StateUpdate

Reviewed By: boryas

Differential Revision: D15709626

fbshipit-source-id: a50b616ca6e57c0542b582c3e721ce11e2aca969