Enable Algorithmic LPM for stand-alone RIB
Summary:
Broadcom XGS devices require a default route to always be present when
they are configured in Algorithmic Longest Prefix Match (ALPM) mode. To satisfy
this requirement, a default route (for each address family) is unconditionally
injected into the RIB during config application. This default route will be the
first to be pushed down into the ASIC because 0.0.0.0/0 and ::/0 are the
smallest elements in the ordering induced on prefixes by RoutePrefix's
inequality operator.

Previously, default routes were injected into the SwitchState RIB prior to
config application. Because VRFs are defined in configuration, ALPM only worked
with a single VRF- the default VRF of 0.

Reviewed By: tfangit

Differential Revision: D15269474

fbshipit-source-id: 8832ccd81b1195e76ea07a804832965420c41ab4