Get rid of a gratuitous check purported to prevent a busy loop
Summary:
When a switch state cannot be applied to h/w, we queue
it back into the pendingUpdates list in the hopes of applying it
in a subsequent iteration. Since the update thread's sole purpose
is to dequeue things from pendingUpdates list and try to apply them,
If you think about this mechanism as running in a loop, the above
failed update would create a busy loop (HW would again reject the
update, we would again queue it and so on).

However the handling of pendingUpdates does not run in a loop (timed
or otherwise). It gets triggered when SwSwitch::updateState gets
called for handling a new update. Hence there is no need for the
above check.

Reviewed By: boryas

Differential Revision: D15755112

fbshipit-source-id: 92acc0edd2208f187b09ddd296d0e2e268b6e559