fix race between callbacks and inital state application on warm boot
Summary:
During BcmSwitch::init() we were previously enabling the RX and
linkscan callbacks before re-creating the state of BcmSwitch from the recovered
SwitchState. As a result, we were vulnerable to the following types of races:
* receive a link scan callback (DOWN) while portAndEgressIdsMap is not yet set
up. Don't shrink ecmp groups and effectively blackhole until neighbor expiry
state change.
* receive an rx packet while re-creating the trunk table and have a data race
on trunks_. We don't currently warm boot with trunks, so this one is less of
a concern.
* same as above for sflow exporters

The solution is pretty simple: we simply defer registering the bcm callbacks
until after we have the BcmSwitch in a consistent internal state. There is
no harm in this for rx, because SwSwitch will ignore the packets at this
point anyway. For link scan, the additional link traversal ensures we don't
miss any events, so deferring it to this point is safe as well.

Differential Revision: D15763445

fbshipit-source-id: c96e4c8e57865718396dbb127d91ca324e93b035