fix shutdown race in NeighborUpdater
Summary:
If a NeighborCacheEntry AsyncTimeout lambda is in flight when
we delete the NeighborUpdater, then cancelTimeout will not actually take
effect, and we can have a race where we delete an entries cache_ while
still running the lambda.

The fix is to ensure that we drain those cancels in the evb, even if
all the entries are flushed. More details in an inline comment.

Reviewed By: aeckert

Differential Revision: D15782382

fbshipit-source-id: 00dee0969a3fbd7559e64ca99f6db1ba70e6325d