A Stand-Alone Modular RIB
Summary:
The south-bound interface of the stand-alone RIB consists of FIB
programming. Previously, this interface was written with the assumption that the
FIB being programmed is that of FBOSS (namely `ForwardingInformationBase`). As a
necessary consequence, the stand-alone RIB could not be used outside of
FBOSS. The changes herein eliminate this limitation by 1) allowing for any FIB
to be plugged into the stand-alone RIB, 2) removing compilation dependencies on
FBOSS' FIB.

(1) is accomplished by the replacement of
`RoutingInformationBase::ApplyStateUpdateFunction` with
`RoutingInformationBase::FibUpdateFunction`. This user-supplied callback is
invoked after every RIB update, including that caused by config application. The
callback can execute arbitrary FIB programming in accordance with supplied RIB
state.

(2) is accomplished by the introduction of the `fib_updater` library. This
library contains the logic to implement the callback in (1) for FBOSS' fib. It
consists of `ForwardingInformationBaseUpdater` and utility functions to
translate stand-alone RIB objects to FBOSS FIB objects (the latter of which were
previously methods on the stand-alone RIB objects which they translated).

Differential Revision: D15643208

fbshipit-source-id: 0f551612d539457a938802ad19b0adcc82a5e88b