Red diff MockableHwSwitch/Platform
Summary:
We built this to be able to play switch states used by ctrl-test
on actual h/w. However since this was added a few years ago we haven't
really used this to run things on actual h/w. Meanwhile we have pivoted
to bcm_tests where we serve the similar use case, and inspect the
HwSwitch, SDK state etc.
So killing the MockableHwSwitch/Platform.

Reviewed By: aeckert

Differential Revision: D15839251

fbshipit-source-id: 0e373c063fd7d31e01bd7a63dc4eb6a5038e9ab0