Break dependency on HwTestHandle in base Platform class
Summary:
HwTestHandle is only needed in conjunction with
a MockPlatform. So move the api down to that layer rather
than at the base Platform class.
Having it at the base platform class
a) Pollutes the API
b) Makes it harder to break things into libs. In my current
use case I was trying to break apart a BcmPlatform lib, but
due to this dependency, BcmPlatform would ending up pulling
in all of HwTestHandle dependency (which in turn has a pretty
big dependency on SwSwitch)

Reviewed By: phshaikh

Differential Revision: D15911498

fbshipit-source-id: dd78b8a53fb447b6ccf762509a83985278ca2bd8