add back spin lock protecting pending updates list
Summary:
This lock should not have been removed when changing what we
do with failed state updates and makes this block access the pending
updates list w/out any protection, definitely causing a race.

We are seeing segfaults trying to print the name of a state update,
which seems likely to be related to this.

Differential Revision: D15982726

fbshipit-source-id: a92ae335b60ddbb17f6c7b360e62f80e064ce556