fix RestartTimeTracker if no saved state found
Summary:
There was a bug in logic s.t if we don't recover the saved
timestamps on the current restart, we don't save them on the
subsequent one.

This can lead to us never getting the full set of restart timers
populated :(.

This diff changes the logic so that we set the 'completed_' var
properly after FIB_SYNCED, even if we didn't recover the initial
events.

Differential Revision: D16069200

fbshipit-source-id: 99aabdd7c4a733bce94c04e3c7abd5d18024042a