Let the bcm SDK determine trunk IDs
Summary:
Instead of trying to sequentially allocate chip IDs, we can just let the
underlying hardware determine a free ID and use it. This simplifies the
tracking we have to do and facilitiates the warmboot procedure by avoiding
potential conflicts between IDs already in use and IDs for newly created
trunks.

Reviewed By: aeckert

Differential Revision: D16113787

fbshipit-source-id: 6d0657dcc75de5bd7618615a6543079264031f3e