populate label stack to tunnel map
Summary:
using BCM MPLS APIs to populate map of label stack to tunnel id.

This is done only for labeled egress.

If egress is labeled, then it would be used for
1) swap
2) push

In former case, it has no tunnel associated with it.

In latter case, it has tunnel associated with it.

Differential Revision: D16040612

fbshipit-source-id: f5c6101db109682ad30924c9d2479656dafb7aa6