add to-mirror sampling to switch config
Summary:
The sFlow-related fields of the Port config are extracted to a single struct, "PortSflowConfig". An additional field is added to "PortSflowConfig" that allows for configuration of a destination to send sampled packets. This can be the CPU or Mirror destinations.

If the sample destination is set to Mirror, egress sampling must be set to 0 because egress sampling to a mirror destination is not currently supported.

The other changes are merely for the migration of the sflow sample rates to a new struct.

Reviewed By: phshaikh

Differential Revision: D15988109

fbshipit-source-id: b4b377b22d390bff3542074493a1da1806642984