Add support to name queue configs
Summary:
This patch is a step towards making queue config overrides easy. In the current
configuration, SwitchConfig::defaultPortQueues applies to all ports unless a
port overrides the config. However, if a subset of ports (e.g. donwlinks)
requires same override, now the override config must repeat in every downlink.

This patch avoids it by defining a SwitchConfig::portQueueConfigs map from
PortQueueConfigName to PortQueue config. The ports that desire override, can
simply reference the desired config with PortQueueConfigName.

Subsequent diffs in the stack make use of this in coop to generate the config
and in agent to consume it.

Reviewed By: aeckert

Differential Revision: D16236851

fbshipit-source-id: 7c62e4dff35393011d6fa4cad7231e68b64a1727