Support name based per port override for queue configs
Summary:
This patch continues to build on prior diffs in the stack to provide per port
override for queue configs.

If a port's config specifies portQueueConfigName, portQueueConfigName must be
present in portQueueConfigs map or else we throw an error.

Note that agent no longer uses thrift field: portConf->queues. Thus, all the BCM tests
are fixed to use the new portQueueConfigName mechanism.

Subsequent diff in the stack (which will only be landed after an agent with
_this_ patch has rolled across entire fleet) removes queues field from port
config altogether.

Differential Revision: D16236850

fbshipit-source-id: 34030d8d49e79fb7a9945e6b757eef5938779ae5