Don't blindly ignore return value in traversal function
Summary:
We were ignoring traversal function errors while
populating warm boot cache. This meant that if for whatever
reason such a function failed, we would continue with a
inconsistent picture of h/w. Later when we tried to
program a object which we failed to retrieve during traversal
we will endup throwing a exception. However the problem
occurred earlier. So move to catching these at the site of
original error.
Further, its not as if we were consistently ignoring all
return values while populating warm boot cache. Code
for populating ACLs, qos maps, rtag7 settings was already
checking for return value.

Reviewed By: phshaikh

Differential Revision: D16296464

fbshipit-source-id: 4deec14516fdca37b991694f44092f48f321302f