Move LoadBalancerUtils to be under hw/test
Summary:
As titled. This is looking forward to be able to use
this with any HwSwitch implementation. For now since we only
build packet generation, load balancing tests in non-oss targets,
I am only building closed source versions of these libs. Building
OSS versions would be straightforward and will be done as needed

Reviewed By: phshaikh

Differential Revision: D16024615

fbshipit-source-id: 6971a36a49b6acd1500b0b79654843eddc9e18fb