add sflow/gre tunneling choice to switch config
Summary:
Add a "tunnel" field to the Mirror config which allows for the specification between sFlow and GRE tunneling. Within the MirrorTunnel config are optional tunnel fields such as source/dest UDP ports (for sFlow) and ttl. More optional fields can be added as needed.

If a destination IP is provided (ie, non-SPAN mirroring) then a tunnel type is needed. If it isn't specified, or if a switch warmboots from a config that doesn't have the tunnelType field, then the tunnel type currently defaults to ERSPAN.

A tunnel is only optionally specified because SPAN mirroring doesn't require one.

Mirror tests were changed slightly to make buck happy!

Reviewed By: boryas

Differential Revision: D16097805

fbshipit-source-id: 0382d0bb97aa801a67305832eb5125ec64f4b001