warning for non-existing port
Summary: Certain asics can be up without any ports and the ports can be added through the config. During the hw init, an empty port list returned should be acceptable. Changing the exception to warning if the ports exist in the config but not yet created in sw and hw.

Differential Revision: D15787932

fbshipit-source-id: d6abe5b299885d6ff0b960b1fc04e609096fd6cf