add to-mirror sampling to SwState Layer
Summary:
The thrift config applier reads the Sample Destination (which directs sampling to Mirror or CPU) from the switch config and adds it to a Port object. sampleDest is added to PortFields and is stored/restored via folly (which uses to/from json, which uses to/from thrift).

Some small changes to tests to make buck happy.

Reviewed By: phshaikh

Differential Revision: D15991021

fbshipit-source-id: 25a29f619504de035a6ddaa8b6fcf30188294171