add sflow/gre tunneling choice to SwState layer
Summary:
Added a MirrorTunnelType enum to the config to be passed down the SwState->BCM layers and eventually configure the mirror to use the specified tunnel type.

Also added the tunnelType field to Mirror and MirrorTunnel, and set the field with ApplyThriftConfig.

If a Mirror object is created with a destination IP set but tunnelType is unspecified, tunnelType currently defaults to ERSPAN.

Also many test changes to make buck happy after changing the config file!

Reviewed By: phshaikh

Differential Revision: D16099940

fbshipit-source-id: cdf94026fd565b8f8d2085bb7143b0cdf4e0637a