Get rid of backward compat code for missing neighbor entry state
Summary:
D15010203 added code to handle missing neighbor entry state
in serialized entry state, with a TODO to get rid of it once D15010203
is rolled out. D15010203 is not rolled out everywhere yet, but lets
get the diff ready anyways. I will just wait on landing this till
agent with D15010203 is pushed out

Also kill the test that verifies that we handle missing state gracefully
on warm boot. We now always expect state to be serialized for neighbor
entries.

Reviewed By: boryas, saifhhasan

Differential Revision: D15096190

fbshipit-source-id: 8f68366479a523dafe8235249446efedb4d976ff