add to-mirror sampling to bcm layer
Summary:
An optional "sampleDest_" field is added to BcmPort (BcmPort.h) to specify between sampling to CPU or MIRROR. This field is used to specify the flag in "opennsl_port_control_set" that marks whether packets sampled on the port should be sent to cpu or mirror destinaitons. (done in BcmPort.cpp. Also sampleDest_ -> port control mapping is in BcmPort.cpp)

Because sFlow sampling is not supported on Trident2, a platform-specific check is added to avoid a "feature not supported" error when configuring sampling on a trident 2.

Reviewed By: phshaikh

Differential Revision: D15991051

fbshipit-source-id: 3d21e2a128bba889e6a2796f3a0482a6ab0c2281