Use hw/bcmSwitchEnsemble in both tests and benchmarks
Summary:
The mandate for HwSwitchEnsemble is to provide functionality for
- Setting up hardware
- Applying state deltas, config to Hw
- Inducing link up/down events on hw
- Handling switch link, packet events

Earlier this functionality was embedded inside the Hw/BcmTest class and
hence had to be duplicated for hw/Bcm benchmarks. Abstract it out for
reuse.

A unfortunate side effect of this is that link/packet events no longer
make it to individual tests. This is fine for all but the fake SDK
Sflow tests. In subsequent diff I will build the concept of a HwEventObserverIf
which can be implemented by any class interested in consuming h/w events.
HwSwitchEnsemble will then pass these events to queued observers.

Reviewed By: phshaikh

Differential Revision: D16353417

fbshipit-source-id: eb32005f86eb7dbe554274ddacf91fd02e9a9789