Don't leak non const member objects from const APIs
Summary:
Cleanup interfaces for {Hw, Bcm}Test,
{Hw,Bcm}SwitchEnsemble. Mainly the idea here is to
not leak non const pointers from const APIs. The const
consistency I am going for is - to not provide a modifying
access to members via const APIs.
APIs changed are getHwSwitch and getPlatform, this causes
downstream changes to APIs which were calling non const
apis on HwSwitch and Platform.

SwitchState is a exception to this, in that its constness
is conveyed vi the published API. So make sure the
getProgrammedState() const only returns published switch states

Reviewed By: boryas

Differential Revision: D16367900

fbshipit-source-id: fddad89f7323c1955e77ff2122efce969bd95dee