Add HwSwitchEventObserverIf and pass link, packet events to observers
Summary:
With the HwSwitchEnsemble abstraction one thing that got lost
was the ability of individual tests to observer packet, link etc events.
For all except a single (fakesdk Sflow) test this did not matter. To
handle the aforementioned tests and anticipating this need for other
tests in the future, build the concept of HwSwitchEventObservers
and pass on packet, link events to any queued observer. Note
that I cooked up another interface instead of reusing HwSwitch::Callback
since its API is not quite compatible with having multiple observers -
packetReceived for example assumes ownership of received packets
and that would not play well with multiple observers. To keep the
concept general abstract out a tiny interface

Reviewed By: shri-khare

Differential Revision: D16373134

fbshipit-source-id: 2a1b08a38e420987c8b1c181670fb854d3ea5047