Disable CINT generation after setup() for BCM tests
Summary:
Only enable BcmCinter logging for setup() during BCM tests since any SDK calls
logged afterwards aren't verified. This reduces CINT execution time and allows
tests such as BcmLoadBalancerTests to be used for CINT verification as they
otherwise time out due to a large number of calls to bcm_tx().

This diff removes the dryRun class constant from BcmCinter, meant to optionally
disable the creation of an sdk trace file on initialization, as it was unused
and replaces it with a mechanism to globally enable/disable logging to the file.

Reviewed By: shri-khare

Differential Revision: D16569141

fbshipit-source-id: fa2d9c558768205e0072dedfee4228f084c46d94