Provide a hook for tests to collect additional info on test failure
Summary:
In order to facilitate debugging on hard to reproduce
failures, provide a hook for tests to collect additional info
on test failures.
For an immediate use case, I am looking to debug a hard to repro
Sflow sampling test failure. Looking at the state of h/w counters
should give some clues for next steps. For now I have just overriden
this hook in BcmTest, with the hope being that there will be
some snapshot like information we will need to collect on test failure
and each individual test may be interested in aspects of that snapshot.

TODO: move printDiagCmd outside of BcmSwitch into a utility. This
is a generally useful api that doesn't have much to do with BcmSwitch
as a class

Differential Revision: D16559897

fbshipit-source-id: dd595800d139f1b0918199a5a8dea6f6fecde6a1