Don't have setup return SwitchState when none is needed
Summary:
The pattern for tests is to use applyNewState
to apply changes to HwSwitch. This has been the standard
pattern for a pretty long while. However we were still using
the old pattern of setup functions returning the state
that they programmed, in the hope of then using that
to track programmingState. This adds unnecessary confusion
by conveying 2 ways of doing the same thing where only 1 is
needed. So refactor all setup methods to return nothing
(checked by building bcm_tests with this static_assert plugged
in P76039310).
The static assert was only used temporarily. I did not intend
to enforce this requirement. The requirement from setup function
is that it programs the h/w to desired state. If you already
have a function that does that, you are free to reuse it here
regardless of its return value. The current cleanup was intended
to get rid of historically useful but now gratuitous return
values.

Caveat: Most of these changes are done via sed scripts and then
auto formatting. I did tighten some lambda captures by hand, but
that's not the intention of this diff

Reviewed By: phshaikh

Differential Revision: D16578011

fbshipit-source-id: 92d0b100ab0e5c97bd9665422dac5c5123d2c546