add SampleDestination (CPU/Mirror) to Warmboot cache's MirroredPort2Handle map
Summary:
A mirror destination is specified not only by the port and direction (ingress vs egress), but also by whether the destination was configured with an SFLOW flag. If the sampling destination on a port is set to MIRROR, the sampled packets will be sent to all mirrors on that port that were configured with the SFLOW flag.

Accordingly, we add the SampleDestination of a port to its MirroredPort2Handle mapping in the warmboot cache so we can set/get the mirror destination using the correct flags for egress/ingress and for sflow/no sflow.

Reviewed By: phshaikh

Differential Revision: D16546074

fbshipit-source-id: bbacee2332cf8b444a658397e397e008f73f1e68