Enable L2 mac address aging to 5mins
Summary:
See discussion here:

https://fb.workplace.com/groups/netstigations/permalink/2691669807548284/

but our L2 MAC address tables were found to be full in ~100 switches around
the fleet.  Why they were full is a different issue, but it exposed that
we did not have MAC address aging enabled.

This patch enables it and provides a BRCM test to make sure it works.

This is a very standard behavior in ethernet networks and there could
be problems (e.g., if there was a hardware swap and we couldn't learn the
new l2 addrs) if we don't enable it.

Reviewed By: shri-khare

Differential Revision: D16769448

fbshipit-source-id: 3145d42aea58a73d47e615d9af587c3a78c3997d