Disable host route optimization in bcm test for TH, TH3
Summary:
We want host route optimization status to mimic
that in production. This means the following
wedge100, wedge40: on
FAv3 (TH based): off
Galaxy: on
Minipack, YAMP: off

To faithfully mimic this, we would need to run TH tests in
2 modes, which is expensive. So in interest of time, have
TH run with this optimization off and be able to accommodate
higher route scale tests.

Secondly, don't test FSW scale on TD2 it doesn't fit. We
anyways don't have TD2 running as FSWs.

TODO: We have a pending item of evaluating the value of this
optimization now that we have ALPM

Reviewed By: jhshi

Differential Revision: D16931730

fbshipit-source-id: 5b248f0b1f8f9d1d814a01e7110852bede53e735