Assert that a non null passed in switch state was applied exactly
Summary:
BcmSwitch has a overflow protection mechanism for
routes and host entries. For all our current tests we want
the state to apply exactly and not partially (as would be the
case with overflow). So assert that. In the future when
we have overflow tests applyNewState can be easily tweaked
by passing in a boolean which allows for partial state update.
Only overflow tests would be passing that in.

As a concrete use case - we recently added route scale tests
- check that the maximum expected route scale is satisfied
for each platform. We definitely don't want an overflow in
those test

Reviewed By: phshaikh

Differential Revision: D16911380

fbshipit-source-id: 09dda02526de8bef8697b887488ea06732140bf4