fix stats to use SUM/RATE again
Summary:
This was a side effect of recovering port name a bit
earlier. Before we would always create the stat and then end up
overwriting it from the else clause. I still don't 100% understand why
these are different, my guess is that we end up constructing a
Monotonic counter then creating another one via the copy constructor,
but the export types are not properly propagated. ExportedStatMap code
looks correct to me though so that may be incorrect.

Either way, I tested this and removing the brackets fixes the issue so
sending this out.

Reviewed By: sas

Differential Revision: D16918455

fbshipit-source-id: 2c01fd0b441e7a5639ed81e3a22bc423e25b9cd3