Helper functions and macros for route scale benchmarks
Summary:
As titled. Simple function to benchmark
- Speed of inserting a given route distribution in h/w
- Speed of pruning a given route distribution from h/w

Note that this just measure the speed of programming h/w
via the HwSwitch interface. Once we gain experience, we
may cast a wider net and measure both the speed of both
route resolution (in SwSwitch) and programming these
in h/w. For now starting just with the bottom half

Differential Revision: D16954052

fbshipit-source-id: 8949b3ff568692c86dc7fc26ced8763ab4982049