Handle /0 generation for prefix generator
Summary:
We can't use the logic of shifting max bits - masklen
for 0 masklen. Since that would cause a shift error. So handle
the /0 case as a special case.
Note that the semantics here are that multiple getNext would
generate the same default route. This is similar to other
masklen semantics where we would rotate through the range.
Although it seems better to throw a runtime error on
reaching the end of range. Will fix that in a subsequent diff.

Reviewed By: phshaikh

Differential Revision: D16912662

fbshipit-source-id: 9a32319cbed682596f1c9ea38dd50868a33c406e