better logging around bcm test stages
Summary:
Was debugging some test issues and had a bit of trouble
determining when the assert failed. This was due to two factors:

1. Not knowing which stage we are in
2. using async XLOG which
confusingly interleaves the assert failures from gtest and the
background logging.

This switches to synchronous logging and also adds logs around each
stage of the test from HwTest.cpp to make things easier to track.

Differential Revision: D16989416

fbshipit-source-id: 58dafd38d4b17346074932072c937352fe9a81bd