Move QsfpCache thread creation out of SwSwitch
Summary:
One of the early attempts to fix T44774867 caused processChagnedPorts->
BcmPort::program to be invoked as part of BcmSwitch::init.

However, BcmSwitch::init is 'too early' as QsfpCache::init hasn't had chance to
run yet, and thus the qsfpCache thread is not created. This was solved by
introducing another class viz. AutoInitQsfpCache which inherits from QsfpCache,
but creates the thread in the AutoInitQsfpCache constructor instead of
SwSwitch.

While the T44774867 fix no longer requires moving QsfpCache thread creation, it
is still cleaner design to do so, hence this patch. It is separated from
T44774867 fix though.

Differential Revision: D15541114

fbshipit-source-id: c9bd46a18322f0293a6d69b90e4342b31c60fd8f